src/backend/requirements.txt
# Django packages
Django>=4.0
djangorestframework
django-cors-headers
drf-spectacular
django-rest-framework-docs
djangorestframework-simplejwt

# AI 비동기 처리
celery
redis

# AI packages
pandas
numpy
matplotlib
scikit-learn
lightgbm
xgboost
joblib
Pillow